+++ src/export_results/figures/policy_states.py
@@ -1,9 +1,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
-from process_data.sample_creation_scripts.create_structural_est_sample import (
-    create_policy_state,
-)
+from process_data.structural_sample_scripts.policy_state import create_policy_state
 
 
 def plot_SRA_2007_reform(path_dict):
